requirements.txt
openai
langchain
python-dotenv
ipython
langchain-openai